:recycle: Refactor: dropdown과 컴포넌트 간 데이터 효율 리팩토링 (#48)
